Safari-611.1.10/LayoutTests/media/video-supports-fullscreen-expected.txt


Test that video.webkitSupportsFullscreen returns false unless the setting is enbled.

EVENT(loadedmetadata)
RUN(internals.settings.setFullScreenEnabled(false))
EXPECTED (video.webkitSupportsFullscreen == 'false') OK
RUN(internals.settings.setFullScreenEnabled(true))
EXPECTED (video.webkitSupportsFullscreen == 'true') OK
END OF TEST
